+++ src/main/java/com/rnt/test_passing/converter/QuestionConverter.java
@@ -1,4 +1,4 @@
-package com.rnt.test_passing.util;
+package com.rnt.test_passing.converter;
 
 import com.rnt.test_passing.entity.Option;
 import java.util.ArrayList;
