+++ src/main/resources/application-context.xml
@@ -24,7 +24,7 @@
     <constructor-arg ref="questionRepository"/>
   </bean>
 
-  <bean id="questionConverter" class="com.rnt.test_passing.util.QuestionConverter">
+  <bean id="questionConverter" class="com.rnt.test_passing.converter.QuestionConverter">
   </bean>
 
   <bean id="scanner" class="java.util.Scanner">
